requirements.txt
dj-database-url==0.4.0
Django==1.9.2
django-appconf==1.0.1
django-compressor==2.0
django-forms-bootstrap==3.0.1
django-libsass==0.6
django-pipeline==1.6.6
django-tastypie==0.13.3
gunicorn==19.4.5
libsass==0.10.1
Pillow==3.1.1
psycopg2==2.6.1
python-dateutil==2.5.0
python-mimeparse==1.5.1
rcssmin==1.0.6
requests==2.9.1
rjsmin==1.0.12
six==1.10.0
wheel==0.24.0
whitenoise==2.0.6